perf: optimize Visual Agent chat dialog rendering performance
Extract message rendering into a React.memo'd ChatMessageItem component
so that adding a new message no longer re-renders all previous messages.
Each message item now owns its own expand/collapse state and memoizes
expensive parsing (parseGenDone, parseGenError, extractSizeToken, etc).

Also fix the ResizeObserver useEffect that was depending on [messages],
causing the observer to be torn down and recreated on every new message.
The ResizeObserver already captures DOM height changes from new content,
so it only needs to be created once on mount.

Net effect: ~456 lines removed from the monolithic component, replaced
by a focused 300-line memoized component with stable callback refs.
